begin extracting compliance type defintions from metadata-acquisition module
