+++ wherehows-web/app/constants/metadata-acquisition.ts
@@ -1,14 +1,6 @@
 import Ember from 'ember';
 import { arrayMap } from 'wherehows-web/utils/array';
-
-/**
- * Defines the string values that are allowed for a classification
- */
-enum Classification {
-  Confidential = 'confidential',
-  LimitedDistribution = 'limitedDistribution',
-  HighlyConfidential = 'highlyConfidential'
-}
+import { Classification, nonIdFieldLogicalTypes, NonIdLogicalType } from 'wherehows-web/constants/datasets/compliance';
 
 /**
  * String indicating that the user affirms or ignored a field suggestion
@@ -70,83 +62,6 @@ const idLogicalTypes = ['ID', 'URN', 'REVERSED_URN', 'COMPOSITE_URN'];
  */
 const customIdLogicalTypes = ['CUSTOM_ID'];
 
-// Default mapping of field data types to security classification
-// https://iwww.corp.linkedin.com/wiki/cf/display/DWH/List+of+Metadata+for+Data+Sets
-const nonIdFieldLogicalTypes: INonIdLogicalTypes = {
-  NAME: {
-    classification: Classification.Confidential,
-    displayAs: 'Name'
-  },
-  EMAIL: {
-    classification: Classification.Confidential,
-    displayAs: 'E-mail'
-  },
-  PHONE: {
-    classification: Classification.Confidential,
-    displayAs: 'Phone Number'
-  },
-  ADDRESS: {
-    classification: Classification.Confidential,
-    displayAs: 'Address'
-  },
-  LATITUDE_LONGITUDE: {
-    classification: Classification.Confidential,
-    displayAs: 'Latitude and Longitude'
-  },
-  CITY_STATE_REGION: {
-    classification: Classification.LimitedDistribution,
-    displayAs: 'City, State, Region, etcetera'
-  },
-  IP_ADDRESS: {
-    classification: Classification.Confidential,
-    displayAs: 'IP Address'
-  },
-  FINANCIAL_NUMBER: {
-    classification: Classification.Confidential,
-    displayAs: 'Financial Number'
-  },
-  PAYMENT_INFO: {
-    classification: Classification.HighlyConfidential,
-    displayAs: 'Payment Info'
-  },
-  PASSWORD_CREDENTIAL: {
-    classification: Classification.HighlyConfidential,
-    displayAs: 'Password and Credentials'
-  },
-  AUTHENTICATION_TOKEN: {
-    classification: Classification.HighlyConfidential,
-    displayAs: 'Authentication Token'
-  },
-  MESSAGE: {
-    classification: Classification.HighlyConfidential,
-    displayAs: 'Message'
-  },
-  NATIONAL_ID: {
-    classification: Classification.HighlyConfidential,
-    displayAs: 'National Id'
-  },
-  SOCIAL_NETWORK_ID: {
-    classification: Classification.Confidential,
-    displayAs: 'Social Network Id'
-  },
-  EVENT_TIME: {
-    classification: Classification.LimitedDistribution,
-    displayAs: 'Event Time'
-  },
-  TRANSACTION_TIME: {
-    classification: Classification.LimitedDistribution,
-    displayAs: 'Transaction Time'
-  },
-  COOKIE_BEACON_BROWSER_ID: {
-    classification: Classification.Confidential,
-    displayAs: 'Cookies and Beacons and Browser Id'
-  },
-  DEVICE_ID_ADVERTISING_ID: {
-    classification: Classification.Confidential,
-    displayAs: 'Device Id and Advertising Id'
-  }
-};
-
 /**
  * List of non Id field data type classifications
  * @type {Array}
@@ -170,7 +85,9 @@ const idFieldDataTypeClassification: { [K: string]: Classification.LimitedDistri
  * Creates a mapping of nonIdFieldLogicalTypes to default classification for that field
  * @type {Object}
  */
-const nonIdFieldDataTypeClassification: { [K: string]: Classification } = Object.keys(nonIdFieldLogicalTypes).reduce(
+const nonIdFieldDataTypeClassification: { [K: string]: Classification } = (<Array<NonIdLogicalType>>Object.keys(
+  nonIdFieldLogicalTypes
+)).reduce(
   (classification, logicalType) =>
     Object.assign(classification, {
       [logicalType]: nonIdFieldLogicalTypes[logicalType].classification
